Move the code to lrc encoder because only convert to the lrc file need the time-tag based dictionary.
