+++ osu.Game.Rulesets.Karaoke/Utils/TimeTagsUtils.cs
@@ -266,33 +266,6 @@ namespace osu.Game.Rulesets.Karaoke.Utils
 
         /// <summary>
         /// Convert list of time tag to dictionary.
-        /// todo: this method will be removed eventually.
-        /// </summary>
-        /// <param name="timeTags">Time tags</param>
-        /// <param name="applyFix">Should auto-fix or not</param>
-        /// <param name="other">Fix way</param>
-        /// <param name="self">Fix way</param>
-        /// <returns>Time tags with dictionary format.</returns>
-        public static IReadOnlyDictionary<TextIndex, double> ToDictionary(IList<TimeTag> timeTags, bool applyFix = true, GroupCheck other = GroupCheck.Asc,
-                                                                          SelfCheck self = SelfCheck.BasedOnStart)
-        {
-            if (timeTags == null)
-                return new Dictionary<TextIndex, double>();
-
-            // sorted value
-            var sortedTimeTags = applyFix ? FixOverlapping(timeTags, other, self) : Sort(timeTags);
-
-            // convert to dictionary, will get start's smallest time and end's largest time.
-            return sortedTimeTags.Where(x => x.Time != null).GroupBy(x => x.Index)
-                                 .Select(x =>
-                                     x.Key.State == TextIndex.IndexState.Start ? x.FirstOrDefault() : x.LastOrDefault())
-                                 .ToDictionary(
-                                     k => k?.Index ?? throw new ArgumentNullException(nameof(k)),
-                                     v => v?.Time ?? throw new ArgumentNullException(nameof(v)));
-        }
-
-        /// <summary>
-        /// Convert list of time tag to dictionary.
         /// WIll sort by the time.
         /// </summary>
         /// <param name="timeTags">Time tags</param>
